requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.1
cloudinary==1.29.0
dj-database-url==1.0.0
Django==4.0.5
django-cloudinary-storage==0.3.0
django-cors-headers==3.13.0
django-heroku==0.3.1
django-rest-framework==0.1.0
djangorestframework==3.13.1
gunicorn==20.1.0
idna==3.3
Pillow==9.2.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
python-decouple==3.6
pytz==2022.1
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.9
whitenoise==6.2.0
